create ticket POST route
