+++ app/templates/ticket/ticket_create.html
@@ -11,8 +11,8 @@
                 <label>Title</label>
                 <input 
                 class="form-control" 
-                type="title" name="title" 
-                placeholder="Title"
+                type="summary" name="summary" 
+                placeholder="Summary"
                 value="{% block postEditTitle %}{% endblock %}" 
                 autofocus=""
                 >
